+++ examples/Migration/CampaignManagement/CreateCompleteCampaignBothApisPhase3.cs
@@ -21,23 +21,31 @@ using System.Linq;
 using System.Web;
 
 using aw = global::Google.Api.Ads.AdWords.v201809;
+using gag__AdGroupAdRotationMode = Google.Ads.GoogleAds.V1.Enums.AdGroupAdRotationModeEnum.Types.AdGroupAdRotationMode;
+using gag__AdGroupStatus = Google.Ads.GoogleAds.V1.Enums.AdGroupStatusEnum.Types.AdGroupStatus;
+using gag__AdvertisingChannelType = Google.Ads.GoogleAds.V1.Enums.AdvertisingChannelTypeEnum.Types.AdvertisingChannelType;
 using gag__BudgetDeliveryMethod = Google.Ads.GoogleAds.V1.Enums.BudgetDeliveryMethodEnum.Types.BudgetDeliveryMethod;
+using gag__CampaignStatus = Google.Ads.GoogleAds.V1.Enums.CampaignStatusEnum.Types.CampaignStatus;
+using gag__NetworkSettings = Google.Ads.GoogleAds.V1.Resources.Campaign.Types.NetworkSettings;
+
+using gagvc = Google.Ads.GoogleAds.V1.Common;
+using gagver = Google.Ads.GoogleAds.V1.Errors;
 using gagvr = Google.Ads.GoogleAds.V1.Resources;
 using gagvs = Google.Ads.GoogleAds.V1.Services;
 
 namespace Google.Ads.GoogleAds.Examples.V1.Migration.CampaignManagement
 {
     /// <summary>
-    /// This code example is the second in a series of code examples that shows how to create
+    /// This code example is the fourth in a series of code examples that shows how to create
     /// a Search campaign using the AdWords API, and then migrate it to the Google Ads API one
-    /// functionality at a time. See Step0.cs through Step5.cs for code examples in various stages
-    /// of migration.
+    /// functionality at a time. See other examples in this directory for code examples in various
+    /// stages of migration.
     ///
-    /// In this code example, the functionality to create campaign budget has been migrated to
-    /// the Google Ads API. The rest of the functionality - creating a Search campaign, ad groups,
-    /// keywords and expanded text ads are done using the AdWords API.
+    /// In this code example, the functionality to create campaign budget, search campaigns and
+    /// ad groups have been migrated to the Google Ads API. The rest of the functionality - 
+    /// creating keywords and expanded text ads are done using the AdWords API.
     /// </summary>
-    public class Step1 : ExampleBase
+    public class CreateCompleteCampaignBothApisPhase3 : ExampleBase
     {
         /// <summary>
         /// Number of ads being added / updated in this code example.
@@ -60,14 +68,14 @@ namespace Google.Ads.GoogleAds.Examples.V1.Migration.CampaignManagement
         {
             get
             {
-                return "This code example is the second in a series of code examples that " +
+                return "This code example is the fourth in a series of code examples that " +
                     "shows how to create a Search campaign using the AdWords API, and then " +
-                    "migrate it to the Google Ads API one functionality at a time. See Step0.cs " +
-                    "through Step5.cs for code examples in various stages of migration. In " +
-                    "this code example, the functionality to create campaign budget has been " +
-                    "migrated to the Google Ads API. The rest of the functionality - creating a " +
-                    "Search campaign, ad groups, keywords and expanded text ads are done using " +
-                    "the AdWords API.";
+                    "migrate it to the Google Ads API one functionality at a time. See other " +
+                    "examples in this directory for code examples in various stages of " +
+                    "migration. In this code example, the functionality to create campaign " +
+                    "budget, search campaigns and ad groups have been migrated to the Google " +
+                    "Ads API. The rest of the functionality - creating keywords and expanded " +
+                    "text ads are done using the AdWords API.";
             }
         }
 
@@ -82,12 +90,11 @@ namespace Google.Ads.GoogleAds.Examples.V1.Migration.CampaignManagement
             // be used interchangeably.
             long customerId = long.Parse((user.Config as AdWordsAppConfig).ClientCustomerId);
 
-            gagvr::CampaignBudget budget = CreateBudget(client,
-                long.Parse((user.Config as AdWordsAppConfig).ClientCustomerId));
-            aw::Campaign campaign = CreateCampaign(user, budget.Id.Value);
-            aw::AdGroup adGroup = CreateAdGroup(user, campaign.id);
-            aw::AdGroupAd[] adGroupAds = CreateTextAds(user, adGroup.id);
-            aw::AdGroupCriterion[] adGroupCriteria = CreateKeywords(user, adGroup.id,
+            gagvr::CampaignBudget budget = CreateBudget(client, customerId);
+            gagvr::Campaign campaign = CreateCampaign(client, customerId, budget.Id.Value);
+            gagvr::AdGroup adGroup = CreateAdGroup(client, customerId, campaign.Id.Value);
+            aw::AdGroupAd[] adGroupAds = CreateTextAds(user, adGroup.Id.Value);
+            aw::AdGroupCriterion[] adGroupCriteria = CreateKeywords(user, adGroup.Id.Value,
                 KEYWORDS_TO_ADD);
         }
 
@@ -164,140 +171,171 @@ namespace Google.Ads.GoogleAds.Examples.V1.Migration.CampaignManagement
         /// <summary>
         /// Creates the campaign.
         /// </summary>
-        /// <param name="user">The Google Ads user.</param>
+        /// <param name="client">The Google Ads client.</param>
+        /// <param name="customerId">The Google Ads customer ID for which the call is made.</param>
         /// <param name="budgetId">The budget ID.</param>
         /// <returns>The newly created campaign instance.</returns>
-        public aw::Campaign CreateCampaign(AdWordsUser user, long budgetId)
+        private static gagvr::Campaign CreateCampaign(GoogleAdsClient client, long customerId,
+            long budgetId)
         {
             // Get the CampaignService.
-            using (aw::CampaignService campaignService =
-                (aw::CampaignService) user.GetService(AdWordsService.v201809.CampaignService))
+            gagvs::CampaignServiceClient campaignService = client.GetService(
+                Services.V1.CampaignService);
+
+            // Create a budget to be used for the campaign.
+            string budgetResource = gagver::ResourceNames.CampaignBudget(customerId, budgetId);
+
+            // Create the campaign.
+            gagvr::Campaign campaign = new gagvr::Campaign();
+            campaign.Name = "Interplanetary Cruise #" + ExampleUtilities.GetRandomString();
+            campaign.AdvertisingChannelType = gag__AdvertisingChannelType.Search;
+
+            // Recommendation: Set the campaign to PAUSED when creating it to prevent
+            // the ads from immediately serving. Set to ENABLED once you've added
+            // targeting and the ads are ready to serve
+            campaign.Status = gag__CampaignStatus.Paused;
+
+            // Set the bidding strategy and budget.
+            campaign.ManualCpc = new gagvc::ManualCpc();
+            campaign.CampaignBudget = budgetResource;
+
+            // Set the campaign network options.
+            campaign.NetworkSettings = new gag__NetworkSettings
             {
-                // Create the campaign.
-                aw::Campaign campaign = new aw::Campaign
-                {
-                    name = "Interplanetary Cruise #" + ExampleUtilities.GetRandomString(),
-                    advertisingChannelType = aw::AdvertisingChannelType.SEARCH,
+                TargetGoogleSearch = true,
+                TargetSearchNetwork = true,
+                TargetContentNetwork = false,
+                TargetPartnerSearchNetwork = false
+            };
 
-                    // Recommendation: Set the campaign to PAUSED when creating it to prevent
-                    // the ads from immediately serving. Set to ENABLED once you've added
-                    // targeting and the ads are ready to serve.
-                    status = aw::CampaignStatus.PAUSED
-                };
+            // Optional: Set the start date.
+            campaign.StartDate = DateTime.Now.AddDays(1).ToString("yyyyMMdd");
 
-                aw::BiddingStrategyConfiguration biddingConfig =
-                    new aw::BiddingStrategyConfiguration
-                    {
-                        biddingStrategyType = aw::BiddingStrategyType.MANUAL_CPC
-                    };
-                campaign.biddingStrategyConfiguration = biddingConfig;
+            // Optional: Set the end date.
+            campaign.EndDate = DateTime.Now.AddYears(1).ToString("yyyyMMdd");
 
-                campaign.budget = new aw::Budget
-                {
-                    budgetId = budgetId
-                };
+            // Create the operation.
+            gagvs::CampaignOperation operation = new gagvs::CampaignOperation();
+            operation.Create = campaign;
 
-                // Set the campaign network options.
-                campaign.networkSetting = new aw::NetworkSetting
-                {
-                    targetGoogleSearch = true,
-                    targetSearchNetwork = true,
-                    targetContentNetwork = false,
-                    targetPartnerSearchNetwork = false
-                };
+            // Add the campaign.
+            gagvs::MutateCampaignsResponse retVal = campaignService.MutateCampaigns(
+                customerId.ToString(), new gagvs::CampaignOperation[] { operation });
 
-                // Optional: Set the start date.
-                campaign.startDate = DateTime.Now.AddDays(1).ToString("yyyyMMdd");
+            // Retrieve the newly created campaign.
+            string newResourceName = retVal.Results.First().ResourceName;
+            gagvr::Campaign newCampaign = GetCampaign(client, customerId, newResourceName);
 
-                // Optional: Set the end date.
-                campaign.endDate = DateTime.Now.AddYears(1).ToString("yyyyMMdd");
+            // Display the results.
+            Console.WriteLine($"Campaign with ID={newCampaign.Id} and name=" +
+            $"'{newCampaign.Name}' was created.");
 
-                // Create the operation.
-                aw::CampaignOperation operation = new aw::CampaignOperation
-                {
-                    @operator = aw::Operator.ADD,
-                    operand = campaign
-                };
+            // Return the newly created campaign.
+            return newCampaign;
+        }
 
-                // Add the campaign.
-                aw::CampaignReturnValue retVal = campaignService.mutate(
-                    new aw::CampaignOperation[] { operation });
+        /// <summary>
+        /// Gets the campaign.
+        /// </summary>
+        /// <param name="client">The Google Ads client.</param>
+        /// <param name="customerId">The Google Ads customer ID for which the call is made.</param>
+        /// <param name="campaignResourceName">The campaign resource.</param>
+        /// <returns>The campaign.</returns>
+        private static gagvr::Campaign GetCampaign(GoogleAdsClient client, long customerId,
+            string campaignResourceName)
+        {
+            // Get the GoogleAdsService.
+            gagvs::GoogleAdsServiceClient googleAdsService = client.GetService(
+                Services.V1.GoogleAdsService);
 
-                // Retrieve the newly created campaign.
-                aw::Campaign newCampaign = retVal.value[0];
+            // Create the search query.
+            string searchQuery = "SELECT campaign.id, campaign.name, " +
+                $"campaign.resource_name from campaign where " +
+                $"campaign.resource_name='{campaignResourceName}'";
 
-                // Display the results.
-                Console.WriteLine($"Campaign with ID={newCampaign.id} and name=" +
-                    $"'{newCampaign.name}' was created.");
+            // Retrieve the campaigns.
+            PagedEnumerable<gagvs::SearchGoogleAdsResponse, gagvs::GoogleAdsRow>
+                searchPagedResponse = googleAdsService.Search(customerId.ToString(), searchQuery);
 
-                // return the newly created campaign.
-                return newCampaign;
-            }
+            // Return the results.
+            return searchPagedResponse.First().Campaign;
         }
 
         /// <summary>
         /// Creates the ad group.
         /// </summary>
-        /// <param name="user">The Google Ads user.</param>
+        /// <param name="client">The Google Ads client.</param>
+        /// <param name="customerId">The Google Ads customer ID for which the call is made.</param>
         /// <param name="campaignId">The campaign ID.</param>
         /// <returns>The newly created ad group instance.</returns>
-        public aw::AdGroup CreateAdGroup(AdWordsUser user, long campaignId)
+        private static gagvr::AdGroup CreateAdGroup(GoogleAdsClient client, long customerId,
+            long campaignId)
         {
             // Get the AdGroupService.
-            using (aw::AdGroupService adGroupService =
-                (aw::AdGroupService) user.GetService(AdWordsService.v201809.AdGroupService))
+            gagvs::AdGroupServiceClient adGroupService = client.GetService(
+                Services.V1.AdGroupService);
+
+            // Create the ad group.
+            gagvr::AdGroup adGroup = new gagvr::AdGroup()
             {
-                // Create the ad group.
-                aw::AdGroup adGroup = new aw::AdGroup
-                {
-                    name = string.Format("Earth to Mars Cruises #{0}",
-                        ExampleUtilities.GetRandomString()),
-                    status = aw::AdGroupStatus.ENABLED,
-                    campaignId = campaignId,
+                Name = $"Earth to Mars Cruises #{ExampleUtilities.GetRandomString()}",
+                Status = gag__AdGroupStatus.Enabled,
+                Campaign = gagver::ResourceNames.Campaign(customerId, campaignId),
 
-                    // Set the ad group bids.
-                    biddingStrategyConfiguration = new aw::BiddingStrategyConfiguration()
-                    {
-                        bids = new aw::Bids[]
-                        {
-                            new aw::CpcBid
-                            {
-                                bid = new aw::Money
-                                {
-                                    microAmount = 10000000
-                                }
-                            }
-                        }
-                    },
+                // Set the ad group bids.
+                CpcBidMicros = 10000000,
 
-                    // Optional: Set the rotation mode.
-                    adGroupAdRotationMode = new aw::AdGroupAdRotationMode
-                    {
-                        adRotationMode = aw::AdRotationMode.OPTIMIZE
-                    }
-                };
+                // Optional: Set the rotation mode.
+                AdRotationMode = gag__AdGroupAdRotationMode.Optimize
+            };
 
-                // Create the operation.
-                aw::AdGroupOperation operation = new aw::AdGroupOperation
-                {
-                    @operator = aw::Operator.ADD,
-                    operand = adGroup
-                };
+            // Create the operation.
+            gagvs::AdGroupOperation operation = new gagvs::AdGroupOperation()
+            {
+                Create = adGroup
+            };
 
-                // Create the ad group.
-                aw::AdGroupReturnValue retVal = adGroupService.mutate(
-                    new aw::AdGroupOperation[] { operation });
+            // Create the ad groups.
+            gagvs::MutateAdGroupsResponse response = adGroupService.MutateAdGroups(
+                customerId.ToString(), new gagvs::AdGroupOperation[] { operation });
 
-                // Retrieve the newly created ad group.
-                aw::AdGroup newAdGroup = retVal.value[0];
+            // Retrieve the newly created ad group.
+            string newResourceName = response.Results.First().ResourceName;
+            gagvr::AdGroup newAdGroup = GetAdGroup(client, customerId, newResourceName);
 
-                // Display the results.
-                Console.WriteLine($"Ad Group with ID={newAdGroup.id} and name=" +
-                    $"'{newAdGroup.name}' was created.");
+            // Display the results.
+            Console.WriteLine($"Ad group with ID={newAdGroup.Id} and name=" +
+                $"'{newAdGroup.Name}' was created.");
 
-                // Return the newly created ad group.
-                return newAdGroup;
-            }
+            // Return the newly created ad group.
+            return newAdGroup;
+        }
+
+        /// <summary>
+        /// Gets the ad group.
+        /// </summary>
+        /// <param name="client">The Google Ads client.</param>
+        /// <param name="customerId">The Google Ads customer ID for which the call is made.</param>
+        /// <param name="adGroupResourceName">The ad group resource.</param>
+        /// <returns>The ad group.</returns>
+        private static gagvr::AdGroup GetAdGroup(GoogleAdsClient client, long customerId,
+            string adGroupResourceName)
+        {
+            // Get the GoogleAdsService.
+            gagvs::GoogleAdsServiceClient googleAdsService = client.GetService(
+                Services.V1.GoogleAdsService);
+
+            // Create the search query.
+            string searchQuery = "SELECT ad_group.id, ad_group.name, " +
+                $"ad_group.resource_name from ad_group where " +
+                $"ad_group.resource_name='{adGroupResourceName}'";
+
+            // Retrieve the ad groups.
+            PagedEnumerable<gagvs::SearchGoogleAdsResponse, gagvs::GoogleAdsRow>
+                searchPagedResponse = googleAdsService.Search(customerId.ToString(), searchQuery);
+
+            // Return the results.
+            return searchPagedResponse.First().AdGroup;
         }
 
         /// <summary>
